requirements.txt
dbt-bigquery==1.1.0
google-cloud-bigquery==2.34.3
pandas==1.4.2
